pi1_app/server/requirements.txt
flask==3.0.3
paho-mqtt==1.6.1
influxdb-client==1.45.0
